Massive rework of data retrieval.
Data is now retrieved and stored as an ordered dictionary to guarantee both reproducibility and quickness with constant time access.

This patch also introduces a validation ratio to choose from the available validation samples, with the introduction of the `ratio` type for arguments parsing.

More information is also logged upon script launch to convey these changes.